requirements.txt
accelerate==1.9.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.2
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
faiss-cpu==1.11.0.post1
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
gradio==3.50.2
gradio_client==0.6.1
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.9
langsmith==0.4.10
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.5
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==2.0.1
networkx==3.5
numpy==1.26.4
orjson==3.11.1
packaging==25.0
pandas==2.3.1
pillow==10.4.0
propcache==0.3.2
psutil==7.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.26.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
semantic-version==2.10.0
sentence-transformers==5.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.2.2
tqdm==4.67.1
transformers==4.54.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==11.0.3
yarl==1.20.1
zstandard==0.23.0
